+++ src/main/kotlin/fr/insalyon/pld/agile/model/Duration.kt
@@ -6,13 +6,13 @@ import fr.insalyon.pld.agile.lib.graph.model.Measurable
  * A duration between two events
  */
 class Duration private constructor(
-    private val _seconds: Long = 0L
+    private val _seconds: Int = 0
 ) : Comparable<Measurable>, Measurable {
 
-  override val length: Long
+  override val length: Int
     get() = _seconds
 
-  constructor(hour: Int = 0, minutes: Int = 0, seconds: Int = 0): this(hour * 3600L + minutes * 60L + seconds)
+  constructor(hour: Int = 0, minutes: Int = 0, seconds: Int = 0): this(hour * 3600 + minutes * 60 + seconds)
 
   operator fun plus(other: Duration): Duration = Duration(_seconds + other.toSeconds())
 
@@ -42,18 +42,18 @@ class Duration private constructor(
 
   override fun toString():String{
     var secondes = _seconds
-    if(secondes==0L)
+    if(secondes==0)
       return "0s"
-    val hours : Long = secondes / 3600L
-    secondes = secondes - hours * 3600L
-    val minutes : Long = secondes / 60L
-    secondes = secondes - minutes * 60L
+    val hours : Int = secondes / 3600
+    secondes = secondes - hours * 3600
+    val minutes : Int = secondes / 60
+    secondes = secondes - minutes * 60
     var res = ""
-    if(hours != 0L)
+    if(hours != 0)
       res += ""+ hours + "h"
-    if(minutes != 0L)
+    if(minutes != 0)
       res += ""+ minutes + "m"
-    if(secondes != 0L)
+    if(secondes != 0)
       res += ""+ secondes + "s"
     return res
   }
